update: start vxlan mode
