+++ struct/plugins/cniPlugin.go
@@ -1,0 +1,26 @@
+package plugins
+
+import "github.com/containernetworking/cni/pkg/types"
+
+type PluginConf struct {
+	// NetConf 里头指定了一个 plugin 的最基本的信息, 比如 CNIVersion, Name, Type 等, 当然还有在 containerd 中塞进来的 PrevResult
+	types.NetConf
+
+	// 这个 runtimeConfig 是可以在 /etc/cni/net.d/xxx.conf 中配置一个
+	// 类似 "capabilities": {"xxx": true, "yyy": false} 这样的属性
+	// 表示说要在运行时开启 xxx 的能力, 不开启 yyy 的能力
+	// 然后等容器跑起来之后(或者被拉起来之前)可以直接通过设置环境变量 export CAP_ARGS='{ "xxx": "aaaa", "yyy": "bbbb" }'
+	// 来开启或关闭某些能力
+	// 然后通过 stdin 标准输入读进来的数据中就会多出一个 RuntimeConfig 属性, 里面就是 runtimeConfig: { "xxx": "aaaa" }
+	// 因为 yyy 在 /etc/cni/net.d/xxx.conf 中被设置为了 false
+	// 官方使用范例: https://kubernetes.feisky.xyz/extension/network/cni
+	// cni 源码中实现: /cni/libcni/api.go:injectRuntimeConfig
+	RuntimeConfig *struct {
+		TestConfig map[string]interface{} `json:"testConfig"`
+	} `json:"runtimeConfig"`
+
+	// 这里可以自由定义自己的 plugin 中配置了的参数然后自由处理
+	Bridge string `json:"bridge"`
+	Subnet string `json:"subnet"`
+	Mode   string `json:"mode" default:"host-gw"`
+}
